requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
qrcode[pil]==7.4.2
pyzbar==0.1.9
pillow==10.4.0
email-validator==2.2.0
python-dotenv==1.0.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio>=1.3.0
jinja2>=3.1.2
